refactor: fix some clippy warnings

+++ crates/typst-pdf/src/tags/mod.rs
@@ -895,7 +895,7 @@ impl BBoxCtx {
             rect.max.y.to_f32(),
         )
         .unwrap();
-        Some(BBox::new(page_idx as usize, rect))
+        Some(BBox::new(page_idx, rect))
     }
 }
